% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_calendar.R
\name{re_calendar}
\alias{re_calendar}
\title{plot Calendar heatmap using echarts}
\usage{
re_calendar(data, title = "", theme = "light", width = NULL,
  height = NULL, cellSize = 20, elementId = NULL)
}
\arguments{
\item{data}{a data.frame with two columns,the fisrt column is date the second is value}

\item{title}{the title of the plot}

\item{theme}{the theme of the plot ,it can be c("light","dark"),default is theme="light"}
}
\description{
A R interface to Calendar heatmap in echarts
}
\examples{
library(RECalendar)
data = data.frame(date=Sys.Date()+1:400,n=sample(400))
re_calendar(data,title="Calendar Heatmap")
}
\references{
Baidu Echars : http://www.echartsjs.com
}
